Extract donation call to separate file
